refactor(phone-input): better performance and apply theme
